backend/pyproject.toml
[project]
name = "document-llm-analysis-backend"
version = "1.0.0"
description = "FastAPI backend for Document LLM Analysis"
requires-python = ">=3.12"
dependencies = [
# === Web Framework ===
"fastapi>=0.115.0",
"uvicorn[standard]>=0.32.0",
"pydantic>=2.10.0",
"pydantic-settings>=2.6.0",
"python-multipart>=0.0.18",
"sse-starlette>=2.2.0",
"httpx>=0.28.0",
"slowapi>=0.1.9",
"sqlalchemy>=2.0.0",
"python-jose[cryptography]>=3.3.0",
"passlib[bcrypt]>=1.7.4",
"email-validator>=2.3.0",

# === LLM & Agent ===
"langchain>=1.2.0",
"langchain-core>=1.2.0",
"langchain-community>=0.4.0",
"langchain-openai>=1.1.0",
"langchain-text-splitters>=1.1.0",
"langchain-google-genai>=4.2.0",
"langgraph>=1.0.0",
"langgraph-prebuilt>=1.0.0",
"google-generativeai>=0.8.0",
"openai>=2.14.0",
"tiktoken>=0.12.0",
"instructor>=1.13.0",

# === RAG & Vector DB ===
"chromadb>=1.3.0",
"sentence-transformers>=5.2.0",

# === Data Processing ===
"pandas>=2.3.0",
"numpy>=2.2.0",
"openpyxl>=3.1.0",
"pypdf>=6.5.0",
"python-docx>=1.2.0",
"lxml>=6.0.0",
"markitdown",
"duckdb>=1.0.0",
"markdown>=3.10.0",

# === Statistical Analysis ===
"scipy>=1.16.0",
"scikit-learn>=1.8.0",
"matplotlib>=3.10.0",
"plotly>=6.5.0",
"statsmodels>=0.14.0",

# === Utilities ===
"python-dotenv>=1.2.0",
"python-magic>=0.4.27",
"tenacity>=9.1.0",
"rich>=14.0.0",
"mcp>=1.1.2",
"thefuzz[speedup]",
"structlog>=25.5.0",
"redis>=5.0.0",
"prometheus-client>=0.20.0",
"prometheus-fastapi-instrumentator>=7.0.0",
"tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "bandit>=1.7.0",
]
ml = [
    # 進階 ML 功能（可選）
    "torch>=2.9.0",
    "transformers>=4.57.0",
    "prophet",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
